requirements.txt
requests
transifex-python
